% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{example_lightbox2}
\alias{example_lightbox2}
\title{Run an example \code{\link{lightbox2}} widget}
\usage{
example_lightbox2(...)
}
\arguments{
\item{...}{Named parameters passed to \code{\link{lightbox2}} to test
different stylings.}
}
\description{
Using the stock images in \code{\link{example_images}} launch a
\code{\link{lightbox2}} widget.
}
