% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightbox2.R
\name{lightbox2}
\alias{lightbox2}
\title{Create a lightbox2 gallery of image files}
\usage{
lightbox2(images, titles = images, thumbnailWidth = "100px",
  thumbnailHeight = "100px", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{images}{Vector of paths to image files. Full file paths are
probably recommended.}

\item{titles}{Vector of title/alt names to use for the <img> tag. Should be
the same length as \code{images}.}

\item{thumbnailWidth}{String denoting the width of the images when they are
displayed as thumbnails. Usually in the format "Npx".}

\item{thumbnailHeight}{String denoting the height of the images when they are
displayed as thumbnails. Usually in the format "Npx".}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\description{
Use \href{https://lokeshdhakar.com/projects/lightbox2/}{lightbox2} to display
image files.
}
