% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lightbox2.R
\name{lightbox2}
\alias{lightbox2}
\title{htmlwidget to create a lightbox2 gallery of image files}
\usage{
lightbox2(images, thumbnailImages = NULL, gallery = "lb-gallery",
  titles = NULL, thumbnailAlign = c("center", "left", "right",
  "justify"), thumbnailWidth = "auto", thumbnailHeight = "100px",
  maxHeight = "412px", margin = "auto", searchable = FALSE,
  width = "100\%", height = "auto", elementId = NULL)
}
\arguments{
\item{images, thumbnailImages}{Vector of paths to image files. Full file paths
are recommended. The default is to have the main pictures in \code{images}
be the thumbnail images as well and then control their sizing with
\code{thumbnailWidth} and \code{thumbnailHeight}.}

\item{gallery}{Name of the lightbox gallery. If there are multiple lightbox2
galleries on a single page it might help to name them differently.}

\item{titles}{Optional vector of title/alt names to use for the \code{<img>}
tag. Should be the same length as \code{images}. If none are provided the
file names from \code{thumbnailImages} then \code{images} is used,
whichever is available in that order.}

\item{thumbnailAlign}{String denoting how to arrange the thunbnail images in
the \code{<div>} tag. Gets passed to the \code{<div>}'s text-align style.
Options are "center", "left", "right", "justify".}

\item{thumbnailWidth}{String denoting the width of the images in css units
when they are displayed as thumbnails. The default is \code{"auto"} so the
images are not stretched while their size is controlled by
\code{thumbnailHeight}.}

\item{thumbnailHeight}{String denoting the height of the images in css units
when they are displayed as thumbnails. The default is \code{"100px"}.}

\item{maxHeight}{The max height of the overall widget (the gallery of
thumbnail images) in css units. Any overflow will cause a scroll bar to
appear. The default is \code{"412px"} which allows for 4 rows of images
when \code{thumbnailHeight = "100px"}. 412px and not 400px because each image
gets a bottom (and right) margin of 3px.}

\item{margin}{The margin style used for the overall widget (the gallery of
thumbnail images). Default is "auto" to center the gallery on the page.}

\item{searchable}{Boolean if a search box should be added to filter the
images based on \code{titles}.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{elementId}{Use an explicit element ID for the widget (rather than an
automatically generated one). Useful if you have other JavaScript that
needs to explicitly discover and interact with a specific widget instance.}
}
\description{
Use \href{https://lokeshdhakar.com/projects/lightbox2/}{lightbox2} to display
image files.
}
\examples{
library(lightbox2)
imgs <- example_images() # get some stock image paths
lightbox2(imgs)
}
